heroku_app/requirements.txt
Flask==0.12.2
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
matplotlib==2.0.2
Pillow==4.2.1
requests==2.18.1